+++ app/templates/header.html
@@ -3,10 +3,15 @@
     <button class="navbar-toggler position-absolute d-md-none collapsed" type="button" data-bs-toggle="collapse" data-bs-target="#sidebarMenu" aria-controls="sidebarMenu" aria-expanded="false" aria-label="Toggle navigation">
       <span class="navbar-toggler-icon"></span>
     </button>
-    <form class="w-100" method="POST">
-      <input class="form-control form-control-dark w-100" type="text" name="search" placeholder="Search" aria-label="Search">
-    </form>
-    <div class="navbar-nav">
+    <form class="w-100 d-flex">
+      <input id="search_field" class="form-control form-control-dark w-100" type="text" name="search" placeholder="Search" aria-label="Search">
+      <div id="search_button" class="d-flex justify-content-center alignitems-center">
+        <div class="mt-2 pt-1">
+          <span data-feather="search"></span>
+        </div>
+      </div>
+      </form>
+      <div class="navbar-nav">
         {% if current_user.is_authenticated %}
         <div class="nav-item text-nowrap">
             <a class="nav-link px-3" href="{{ url_for('auth.logout') }}">Sign out</a>
