--- app/templates/search.html
@@ -1,95 +1,0 @@
-{% extends "base.html" %}
-
-{% block links %}
-  <link href="{{ url_for('static', filename='css/user.css') }}" rel="stylesheet" media="screen">
-{% endblock %}
-
-{% block content %}
-<main class="col-md-9 ms-sm-auto col-lg-10 px-md-4">
-
-    <h3 class="mt-1">Search results</h3>
-
-  <div class="d-flex justify-content-between pt-3">
-    <h3 class="mt-1">Users</h3>
-    <a class="btn btn-primary mt-1 add_button" href="#">Filter</a>
-  </div>
-  <div class="table-responsive">
-    <table class="table table-striped table-sm">
-      <thead>
-        <tr>
-          <th scope="col">id</th>
-          <th scope="col">Name</th>
-          <th scope="col">Role</th>
-          <th scope="col">Action</th>
-        </tr>
-      </thead>
-      <tbody>
-          {% for user in users%}
-            <tr>
-              {% if not user.deleted %}
-                  <td>{{ user.id }}</td>
-                  <td><a class="user_update" href="{{ url_for('users.user_update', user_id=user.id) }}">{{ user.username }}</a></td>
-                  <td>{{ user.role.name }}</td>
-                  <td><a class="delete_button" onclick="return confirm('Are you sure?');" href="{{ url_for('users.user_delete', user_id=user.id) }}">Delete</a></td>
-              {% endif %}
-            </tr>
-          {% endfor %}
-      </tbody>
-    </table>
-  </div>
-
-  <div class="d-flex justify-content-between pt-3">
-    <h3 class="mt-1">Accounts</h3>
-    <a class="btn btn-primary mt-1 add_button" href="{{ url_for('accounts.account_add') }}">New Account</a>
-  </div>
-  <div class="table-responsive">
-    <table class="table table-striped table-sm">
-      <thead>
-        <tr>
-          <th scope="col">id</th>
-          <th scope="col">Login</th>
-          <th scope="col">Created By</th>
-          <th scope="col">Created At</th>
-        </tr>
-      </thead>
-      <tbody>
-        {% for account in accounts %}
-          <tr>
-            <td>{{account.id}}</td>
-            <td><a class="user_update" href="{{ url_for('accounts.account_info', account_id=account.id) }}">{{account.login}}</a></td>
-            <td>{{account.user.username}}</td>
-            <td>{{ account.created_at.strftime('%Y-%m-%d, %H:%M') }}</td>
-          </tr>
-        {% endfor %}
-      </tbody>
-    </table>
-  </div>
-
-  <div class="d-flex justify-content-between pt-3">
-    <h3 class="mt-1">Billings</h3>
-    <a class="btn btn-primary mt-1 add_button" href="{{ url_for('billings.billing_add') }}">New Billing</a>
-  </div>
-  <div class="table-responsive">
-    <table class="table table-striped table-sm">
-      <thead>
-        <tr>
-          <th scope="col">id</th>
-          <th scope="col">Credits</th>
-          <th scope="col">Created By</th>
-          <th scope="col">Created At</th>
-        </tr>
-      </thead>
-      <tbody>
-        {% for billing in billings %}
-          <tr>
-            <td>{{billing.id}}</td>
-            <td>{{billing.credits}}</td>
-            <td>{{billing.reseller.username}}</td>
-            <td>{{ billing.created_at.strftime('%Y-%m-%d, %H:%M') }}</td>
-          </tr>
-        {% endfor %}
-      </tbody>
-    </table>
-  </div>
-</main>
-{% endblock %}
